fix updateAUR script; add percentage symbol on polybar battery

+++ scripts/bin/updateAUR.sh
@@ -1,16 +1,18 @@
 #!/usr/bin/bash
 
 PKGS=$(cower -u | awk '{print $2}')
-BUILD_DIR="~/.builds"
+BUILD_DIR="/home/augusto/.builds"
 
 
 for i in $PKGS;
 do
-    echo "Updating: $i"
-    cd ${i}
-    git reset --hard HEAD
-    git clean -d -fx
-    git pull origin master
-    makepkg -si --noconfirm
-    cd ..
+    if [ -d "${BUILD_DIR}/${i}" ]; then
+        echo "Updating: $i"
+        cd ${BUILD_DIR}/${i}
+        git clean -d -fx
+        git reset --hard HEAD
+        git pull origin master
+        makepkg -si --noconfirm
+        cd ..
+    fi
 done
